BVAL-327 Move executable objects to the executable package
