+++ src/main/java/javax/validation/executable/ValidateExecutable.java
@@ -14,7 +14,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package javax.validation;
+package javax.validation.executable;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
